--- djangosaml2idp/models/__init__.py
@@ -1,5 +1,0 @@
-from .ServiceProviderModel import ServiceProvider
-
-__all__ = [
-    "ServiceProvider"
-]
