--- djangosaml2idp/mxins/PermissionRequiredMixin.py
@@ -1,34 +1,0 @@
-from .AccessMixin import AccessMixin
-from django.core.exceptions import ImproperlyConfigured
-
-class PermissionRequiredMixin(AccessMixin):
-    """Verify that the current user has all specified permissions."""
-    permission_required = None
-
-    def get_permission_required(self):
-        """
-        Override this method to override the permission_required attribute.
-        Must return an iterable.
-        """
-        if self.permission_required is None:
-            raise ImproperlyConfigured(
-                '{0} is missing the permission_required attribute. Define {0}.permission_required, or override '
-                '{0}.get_permission_required().'.format(self.__class__.__name__)
-            )
-        if isinstance(self.permission_required, str):
-            perms = (self.permission_required,)
-        else:
-            perms = self.permission_required
-        return perms
-
-    def has_permission(self):
-        """
-        Override this method to customize the way permissions are checked.
-        """
-        perms = self.get_permission_required()
-        return self.request.user.has_perms(perms)
-
-    def dispatch(self, request, *args, **kwargs):
-        if not self.has_permission():
-            return self.handle_no_permission()
-        return super().dispatch(request, *args, **kwargs)
