+++ djangosaml2idp/mxins/__init__.py
@@ -1,7 +1,9 @@
+"""
+mixins
+"""
 from .AccessMixin import AccessMixin
 from .LoginRequiredMixin import LoginRequiredMixin
-from .PermissionRequiredMixin import PermissionRequiredMixin
 from .UserPassesTestMixin import UserPassesTestMixin
 from .IdPHandlerViewMixin import IdPHandlerViewMixin
 
-__all__=["AccessMixin","LoginRequiredMixin","PermissionRequiredMixin","UserPassesTestMixin","IdPHandlerViewMixin"]
+__all__=["AccessMixin","LoginRequiredMixin","UserPassesTestMixin","IdPHandlerViewMixin"]
