--- djangosaml2idp/templates/dev/hook.html
@@ -1,6 +1,0 @@
-正在跳转请稍侯.......
-<script src="http://code.jquery.com/jquery-latest.js"></script>
-<script type="text/javascript">
-    localStorage.setItem("spauthn",'{{spauthn}}');
-    window.location.href = '{{sso_url | safe}}';
-</script>
