--- djangosaml2idp/templates/dev/mock_login.html
@@ -1,34 +1,0 @@
-<!-- <form class='form' method="POST" id="form_login" style="display: none;">
-    username<input name="username">
-    password<input name="password">
-    <button type='button' id='submit'>submit</button>
-</form> -->
-
-<script src="http://code.jquery.com/jquery-latest.js"></script>
-<script type="text/javascript">
-    const urlParams = new URLSearchParams(window.location.search);
-    const SAMLRequest = urlParams.get('SAMLRequest');
-    let next = '{{next}}'+'?SAMLRequest='+SAMLRequest;
-    spauthn = localStorage.getItem("spauthn");
-    console.log(spauthn);
-    if( spauthn!=null && spauthn != "" ){
-        localStorage.removeItem("spauthn");
-        window.location.href = next+'&spauthn='+spauthn;
-    };
-
-    // document.getElementById("form_login").setAttribute("style","display:block");
-
-    // $('#submit').click(function(){
-    //     const username = $('input[name=username]').val();
-    //     const password = $('input[name=password]').val();
-    //     $.ajax({
-    //         method: 'POST',
-    //         url: '{{token_url}}',
-    //         data: {'username': username, 'password': password},
-    //         success: function(data){
-    //             // window.location.href = next+'&spauthn='+data["data"]["token"];
-    //         }
-    //     })
-    //     return false;
-    // })
-</script>
